Only set hard and soft delete flags when process is ended (#190)
Instance delete flags should not be set before the entire process is ended. Currently applications with autodelete and multiple data tasks fails on second data taks.
Fixes #189

+++ src/Altinn.App.Core/Implementation/DefaultTaskEvents.cs
@@ -204,7 +204,7 @@ public class DefaultTaskEvents : ITaskEvents
 
     private async Task RunAutoDeleteOnProcessEnd(Instance instance, Guid instanceGuid)
     {
-        if (_appMetadata.AutoDeleteOnProcessEnd)
+        if (_appMetadata.AutoDeleteOnProcessEnd && instance.Process?.Ended != null)
         {
             int instanceOwnerPartyId = int.Parse(instance.InstanceOwner.PartyId);
             await _instanceClient.DeleteInstance(instanceOwnerPartyId, instanceGuid, true);
